pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "noteagator"
version = "0.1.0"
description = "Notebook Navigator a keyboard first note tool"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "William Lindsey" }]
keywords = ["cli", "notes", "markdown", "productivity"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.2,<9.0",
    "pyyaml>=6,<7",
    "rich>=14,<15",
    "pyperclip>=1.9,<2.0",
]

[project.urls]
Homepage = "https://github.com/willlindsey05/noteagator"
Issues = "https://github.com/willlindsey05/noteagator/issues"

[project.scripts]
ngt = "noteagator.main:main"
noteagator = "noteagator.main:main"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I"]
